+++ attributes/brotli.rb
@@ -1,0 +1,6 @@
+id = 'ngx'
+
+default[id]['brotli']['version'] = 'v0.1.2'
+default[id]['brotli']['url'] = 'https://github.com/eustas/ngx_brotli'
+default[id]['brotli']['checksum'] = \
+  '309af9e96c10e80f1884acea96379980979581adc287ce338f084607bd48c185'