+++ attributes/ct.rb
@@ -1,6 +1,0 @@
-id = 'ngx'
-
-default[id]['ct']['version'] = '1.3.2'
-default[id]['ct']['url'] = 'https://github.com/grahamedgecombe/nginx-ct/archive'
-default[id]['ct']['checksum'] = \
-  'b4ceae549b9dbf84b2e511633982e4efeee0388e3b7a038a8bac555008a77b88'