+++ attributes/default.rb
@@ -2,24 +2,25 @@ id = 'ngx'
 
 default[id]['install'] = 'stable'
 
-default[id]['stable']['version'] = '1.14.0'
+default[id]['stable']['version'] = '1.14.2'
 default[id]['stable']['checksum'] = \
-  '5d15becbf69aba1fe33f8d416d97edd95ea8919ea9ac519eff9bafebb6022cb5'
+  '002d9f6154e331886a2dd4e6065863c9c1cf8291ae97a1255308572c02be9797'
 
-default[id]['mainline']['version'] = '1.15.3'
+default[id]['mainline']['version'] = '1.15.8'
 default[id]['mainline']['checksum'] = \
-  '9391fb91c3e2ebd040a4e3ac2b2f0893deb6232edc30a8e16fcc9c3fa9d6be85'
+  'a8bdafbca87eb99813ae4fcac1ad0875bf725ce19eb265d28268c309b2b40787'
 
 default[id]['with_openssl'] = true
 default[id]['with_http2'] = true
 default[id]['with_ipv6'] = true
-default[id]['with_ct'] = false
 default[id]['with_headers_more'] = false
 default[id]['with_status'] = false
 default[id]['with_realip'] = false
 default[id]['with_geoip2'] = false
 default[id]['with_secure_link'] = false
 default[id]['with_njs'] = false
+default[id]['with_gzip_static'] = false
+default[id]['with_brotli'] = false
 default[id]['with_debug'] = false
 default[id]['with_devel_kit'] = false
 default[id]['with_lua'] = false