+++ attributes/geoip2.rb
@@ -1,9 +1,9 @@
 id = 'ngx'
 
-default[id]['geoip2']['version'] = '3.0'
+default[id]['geoip2']['version'] = '3.2'
 default[id]['geoip2']['url'] = 'https://github.com/leev/ngx_http_geoip2_module/archive'
 default[id]['geoip2']['checksum'] = \
-  '65a191688348a05d8d92b2e7ce9c6eb8cb8322205c34637da582a1205864133d'
+  '15bd1005228cf2c869a6f09e8c41a6aaa6846e4936c473106786ae8ac860fab7'
 
 default[id]['geoip2']['databases']['dir'] = '/etc/chef-nginx/geolite2'
 
@@ -11,10 +11,10 @@ default[id]['geoip2']['databases']['country']['name'] = 'GeoLite2-Country.mmdb'
 default[id]['geoip2']['databases']['country']['url'] = \
   'http://geolite.maxmind.com/download/geoip/database/GeoLite2-Country.tar.gz'
 default[id]['geoip2']['databases']['country']['checksum'] = \
-  'e5acb422d3a09c3ccbd55b98a0bf926e402dd9cd086d9b70f5d5f7524a8f3676'
+  '1b627bd7a575500cbf9e675630ab2c9b1632c3727f3eab29a06e50568341bdfa'
 
 default[id]['geoip2']['databases']['city']['name'] = 'GeoLite2-City.mmdb'
 default[id]['geoip2']['databases']['city']['url'] = \
   'http://geolite.maxmind.com/download/geoip/database/GeoLite2-City.tar.gz'
 default[id]['geoip2']['databases']['city']['checksum'] = \
-  'c68df20de81f738f047d2ff235c51bae09fc0e83e856aa4a4729b482dd420c5c'
+  '589b8603a6cd98cf134f96de24618ba1d986a954e40409a67f398eb1edbf6084'