+++ attributes/njs.rb
@@ -1,6 +1,6 @@
 id = 'ngx'
 
-default[id]['njs']['version'] = '0.2.3'
+default[id]['njs']['version'] = '0.2.7'
 default[id]['njs']['url'] = 'https://github.com/nginx/njs/archive'
 default[id]['njs']['checksum'] = \
-  '4ca46da4bd4df3c1037e1dbcc95d97389db203d5b0112344a21b7ad946d16a15'
+  'e386f391b8dda7e6396418109270b30bd44c507767c67d9f041d45fe3f0348a9'