+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2019-01-27
+### Added
+- Add `ngx_http_gzip_static_module` module.
+- Add `ngx_brotli` module.
+
+### Changed
+- Update nginx mainline version to 1.15.8 and stable version to 1.14.2.
+- Update GeoLite2 database file checksums.
+- Update OpenSSL and njs versions.
+- Change `main_ext` log_format name to `combined_ext`.
+
+### Removed
+- Remove nginx-ct module.
+
 ## [1.1.1] - 2018-09-09
 ### Changed
 - Update nginx mainline version to 1.15.3.